Revert "prepare resource collection from one getArray in basic resource "
