+++ src/Resources/BasicResourceCollection.php
@@ -2,9 +2,29 @@
 
 namespace Behamin\BResources\Resources;
 
-use Behamin\BResources\Traits\CollectionResource;
-
 class BasicResourceCollection extends BasicResource
 {
-    use CollectionResource;
+    public function __construct($collectionResource, $transform = false)
+    {
+        parent::__construct($collectionResource);
+
+        if ($transform) {
+            $this->transformData();
+        } else {
+            $this->data = [
+                'items' => $this['data'],
+                'count' => isset($this->count) ? $this->count : count($this->data)
+            ];
+        }
+    }
+
+    protected function transformData()
+    {
+        $this->data = [
+            'items' => $this['data']->transform(function ($item) {
+                return $this->getArray($item);
+            }),
+            'count' => isset($this->count) ? $this->count : count($this->data)
+        ];
+    }
 }
